Update manager.py
Check if connection is invalidated on clear_connection which is triggered on rollback. 

Otherwise what can happen is a rollback is trying to occur to make the connection valid again, but gets blocked by this clear_connection event. 

conn.connection is connection.connection - > self._revalidate_connection() wont be called when a connection is invalidated eg. "Can't reconnect until transaction is rolled back".